requirements.txt
streamlit>=1.28.0
ollama>=0.3.0
pillow>=10.0.0